meet: move assistant/src/meet → skills/meet-join/daemon
