+++ assistant/src/runtime/routes/__tests__/meet-internal.test.ts
@@ -12,7 +12,7 @@ import { describe, expect, test } from "bun:test";
 
 import type { MeetBotEvent } from "@vellumai/meet-contracts";
 
-import { MeetSessionEventRouter } from "../../../meet/session-event-router.js";
+import { MeetSessionEventRouter } from "../../../../../skills/meet-join/daemon/session-event-router.js";
 import { handleMeetInternalEvents } from "../meet-internal.js";
 
 // ---------------------------------------------------------------------------
