+++ skills/meet-join/daemon/__tests__/event-publisher.test.ts
@@ -17,9 +17,9 @@ import { beforeEach, describe, expect, test } from "bun:test";
 
 import type { MeetBotEvent } from "@vellumai/meet-contracts";
 
-import type { AssistantEvent } from "../../runtime/assistant-event.js";
-import { assistantEventHub } from "../../runtime/assistant-event-hub.js";
-import { DAEMON_INTERNAL_ASSISTANT_ID } from "../../runtime/assistant-scope.js";
+import type { AssistantEvent } from "../../../../assistant/src/runtime/assistant-event.js";
+import { assistantEventHub } from "../../../../assistant/src/runtime/assistant-event-hub.js";
+import { DAEMON_INTERNAL_ASSISTANT_ID } from "../../../../assistant/src/runtime/assistant-scope.js";
 import {
   meetEventDispatcher,
   publishMeetEvent,
